refactor(entity): drop NewChainStore default fallback for fullIDCacheMaxCount
Let the caller supply fullIDCacheMaxCount, matching lruCapacity and
fullCacheDataSizeLimit which have no in-constructor defaults. Remove the
DefaultFullIDCacheMaxCount constant and the non-positive fallback; the
default now lives with the driver flag that populates startup.Config.
The parameter is now uint64 end to end.

+++ driver/controller/startup/entity.go
@@ -21,7 +21,7 @@ func newEntityController(
 	chainID string,
 	storeCacheSize int,
 	storeFullCacheSize int,
-	storeFullIDCacheMaxCount int,
+	storeFullIDCacheMaxCount uint64,
 	monitor persistent.MetricsMonitor,
 ) *entityController {
 	return &entityController{
